refactor(crm): drop Huma suffixes on holder :hammer:
The six exported fiber.Ctx terminals across HolderHandler and
HolderAccountsHandler were unreachable — registerCRMRoutesToApp passes
middleware only and hands the terminals to Huma.

holder_accounts.go was scoped in: GetAccountsByHolder hangs off a different
handler type, but its Huma terminal, its envelope types and
RegisterHolderAccountsRoutes all live in the file being renamed and it
registers inside the holder block of the registrar.

- Delete the dead Fiber wrappers and their tests.
- Port the branches those tests covered onto the Huma transport; every core
  holds at or above baseline and getAccountsByHolder gains (85% -> 100%).
  createHolder stays at its 77.4% baseline.
- Rename methods and envelope types: CreateHolderHuma -> CreateHolder,
  CreateHolderInputHuma -> CreateHolderRequest, and so on.
- holder_handler_huma.go -> holder_handler.go, holder_huma_test.go ->
  holder_handler_test.go.

Three files outside the resource mounted the dead terminals directly and could
not survive the deletion. crm_idempotency_test.go held the holder Fiber
idempotency test plus fakeCRMIdempotencyRepo, which instrument's tests also
use; the test is ported as TestCreateHolder_IdempotentReplay and the shared
fake stays. crm_error_contract_test.go mounted CreateHolder and GetHolderByID
directly; both are ported to the Huma harness. One assertion there changes
shape: the live RFC 9457 envelope carries per-field detail as
errors[]{message,location} (pkg/net/http/problem.go fieldsToErrors), not the
flat fields map the dead Fiber path produced — the codes and statuses are
unchanged. transaction_idempotency_huma_test.go referenced a renamed envelope
type.

No route, operation ID, or wire contract changes: check-docs reports no
drift.

X-Lerian-Ref: 0x1
