+++ components/ledger/internal/adapters/http/in/crm_idempotency_test.go
@@ -5,30 +5,16 @@
 package in
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"io"
-	"net/http/httptest"
-	"testing"
 	"time"
 
-	libConstants "github.com/LerianStudio/lib-commons/v6/commons/constants"
-	"github.com/gofiber/fiber/v3"
-	"github.com/google/uuid"
 	"github.com/redis/go-redis/v9"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/mock/gomock"
-
-	"github.com/LerianStudio/midaz/v4/components/ledger/internal/crm/adapters/mongodb/holder"
-	"github.com/LerianStudio/midaz/v4/components/ledger/internal/crm/services"
-	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
-	"github.com/LerianStudio/midaz/v4/pkg/net/http"
 )
 
 // fakeCRMIdempotencyRepo is an in-memory IdempotencyRepo with SetNX semantics,
-// shared across the two requests in a replay test.
+// shared by the CRM handler tests whose flows claim an idempotency slot. One
+// instance is shared across the requests of a single replay test so the second
+// request sees the first one's claim.
 type fakeCRMIdempotencyRepo struct {
 	store map[string]string
 }
@@ -60,74 +46,4 @@ func (f *fakeCRMIdempotencyRepo) Set(_ context.Context, key, value string, _ tim
 	f.store[key] = value
 
 	return nil
-}
-
-func TestHolderHandler_CreateHolder_IdempotentReplay(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	t.Cleanup(ctrl.Finish)
-
-	orgUUID := uuid.New()
-	orgID := orgUUID.String()
-	holderID := uuid.New()
-	name := "John Doe"
-	document := "91315026015"
-	holderType := "NATURAL_PERSON"
-
-	mockHolderRepo := holder.NewMockRepository(ctrl)
-	// The Mongo create MUST run exactly once across the two identical requests.
-	mockHolderRepo.EXPECT().
-		Create(gomock.Any(), orgID, gomock.Any()).
-		Return(&mmodel.Holder{ID: &holderID, Name: &name, Document: &document, Type: &holderType}, nil).
-		Times(1)
-
-	uc := &services.UseCase{
-		HolderRepo:  mockHolderRepo,
-		Idempotency: newFakeCRMIdempotencyRepo(),
-	}
-	handler := &HolderHandler{Service: uc}
-
-	app := fiber.New()
-	app.Post(
-		"/v2/organizations/:organization_id/holders",
-		func(c fiber.Ctx) error {
-			c.Locals("organization_id", orgUUID)
-			return c.Next()
-		},
-		http.WithBody(new(mmodel.CreateHolderInput), handler.CreateHolder),
-	)
-
-	body := `{"type":"NATURAL_PERSON","name":"John Doe","document":"91315026015"}`
-
-	doRequest := func() (int, string, []byte) {
-		req := httptest.NewRequest(fiber.MethodPost, "/v2/organizations/"+orgID+"/holders", bytes.NewBufferString(body))
-		req.Header.Set("Content-Type", "application/json")
-		req.Header.Set(libConstants.IdempotencyKey, "holder-key-1")
-
-		resp, err := app.Test(req)
-		require.NoError(t, err)
-
-		respBody, err := io.ReadAll(resp.Body)
-		require.NoError(t, err)
-
-		return resp.StatusCode, resp.Header.Get(libConstants.IdempotencyReplayed), respBody
-	}
-
-	// First request: real create, not a replay.
-	status1, replayed1, body1 := doRequest()
-	assert.Equal(t, 201, status1)
-	assert.Equal(t, "false", replayed1)
-
-	var first map[string]any
-	require.NoError(t, json.Unmarshal(body1, &first))
-
-	// Second identical request: replays the original entity, no second create.
-	status2, replayed2, body2 := doRequest()
-	assert.Equal(t, 201, status2)
-	assert.Equal(t, "true", replayed2)
-
-	var second map[string]any
-	require.NoError(t, json.Unmarshal(body2, &second))
-
-	assert.Equal(t, first["id"], second["id"])
-	assert.Equal(t, first["name"], second["name"])
 }
