+++ components/ledger/internal/adapters/http/in/holder.go
@@ -10,17 +10,13 @@ import (
 	"time"
 
 	"github.com/LerianStudio/midaz/v4/components/ledger/internal/crm/services"
-	"github.com/LerianStudio/midaz/v4/pkg"
-	cn "github.com/LerianStudio/midaz/v4/pkg/constant"
 	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
 	"github.com/LerianStudio/midaz/v4/pkg/net/http"
 
 	libCommons "github.com/LerianStudio/lib-commons/v6/commons"
-	libConstants "github.com/LerianStudio/lib-commons/v6/commons/constants"
 	libObservability "github.com/LerianStudio/lib-observability/v2"
 	libLog "github.com/LerianStudio/lib-observability/v2/log"
 	libOpentelemetry "github.com/LerianStudio/lib-observability/v2/tracing"
-	"github.com/gofiber/fiber/v3"
 	"github.com/google/uuid"
 	"go.opentelemetry.io/otel/attribute"
 )
@@ -29,13 +25,20 @@ type HolderHandler struct {
 	Service *services.UseCase
 }
 
-// createHolder is the transport-agnostic core for the holder create. It runs the
-// full idempotency dance (claim + replay-or-create + store) using an
-// already-resolved client key + TTL, so both the Fiber wrapper (CreateHolder) and
-// the Huma shell (CreateHolderHuma) share one implementation and neither touches
-// the other's request/response object. It returns replayed=true when the response
-// was served from a cached idempotency slot so the caller can set the
-// X-Idempotency-Replayed header on its own transport.
+// --- Transport-agnostic cores -------------------------------------------------
+//
+// The createHolder/getHolderByID/... methods below own the span, the service call
+// and — for the create — the idempotency dance. They take primitive args — parsed
+// UUIDs, the decoded payload, the query map — so nothing transport-shaped reaches
+// them; the handlers in holder_handler.go pull those out of the request envelope.
+// Every canonical Midaz error a core returns is rendered by its caller via
+// http.HumaProblem, which fixes the code + HTTP status. Body decode+validation
+// happens BEFORE these cores, in the handler, via http.DecodeAndValidate.
+
+// createHolder owns the span + the full idempotency dance (claim + replay-or-create
+// + store) for an already-decoded payload, using an already-resolved client key +
+// TTL. It returns replayed=true when the response came from a cached idempotency
+// slot, so the handler can set the X-Idempotency-Replayed header.
 func (handler *HolderHandler) createHolder(ctx context.Context, organizationID uuid.UUID, payload *mmodel.CreateHolderInput, clientKey string, ttl time.Duration) (holder *mmodel.Holder, replayed bool, err error) {
 	logger, tracer, reqId, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -97,37 +100,7 @@ func (handler *HolderHandler) createHolder(ctx context.Context, organizationID u
 	return out, false, nil
 }
 
-// CreateHolder is a method that creates Holder information.
-func (handler *HolderHandler) CreateHolder(p any, c fiber.Ctx) error {
-	ctx := c.Context()
-
-	payload, ok := p.(*mmodel.CreateHolderInput)
-	if !ok || payload == nil {
-		return http.WithError(c, pkg.ValidateInternalError(nil, cn.EntityHolder))
-	}
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	clientKey, ttl := http.GetIdempotencyKeyAndTTL(c)
-
-	c.Set(libConstants.IdempotencyReplayed, "false")
-
-	holder, replayed, err := handler.createHolder(ctx, organizationID, payload, clientKey, ttl)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	if replayed {
-		c.Set(libConstants.IdempotencyReplayed, "true")
-	}
-
-	return http.Created(c, holder)
-}
-
-// getHolderByID is the transport-agnostic core for the holder read.
+// getHolderByID reads a single holder by ID.
 func (handler *HolderHandler) getHolderByID(ctx context.Context, organizationID, id uuid.UUID, includeDeleted bool) (*mmodel.Holder, error) {
 	_, tracer, reqId, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -151,34 +124,9 @@ func (handler *HolderHandler) getHolderByID(ctx context.Context, organizationID,
 	return holder, nil
 }
 
-// GetHolderByID retrieves Holder details by a given id
-func (handler *HolderHandler) GetHolderByID(c fiber.Ctx) error {
-	ctx := c.Context()
-
-	id, err := http.GetUUIDFromLocals(c, "id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	includeDeleted := http.GetBooleanParam(c, "include_deleted")
-
-	holder, err := handler.getHolderByID(ctx, organizationID, id, includeDeleted)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, holder)
-}
-
-// updateHolder is the transport-agnostic core for the holder update. fieldsToRemove
-// carries the RFC 7396 merge-patch null-field paths; the Fiber wrapper reads them
-// from the patchRemove local produced by http.WithBody, the Huma shell derives them
-// from the parsed body via http.FindNilFields — both feed this one core.
+// updateHolder applies an already-decoded merge patch. fieldsToRemove carries the
+// RFC 7396 null-field paths the handler derived from the raw body via
+// http.FindNilFields.
 func (handler *HolderHandler) updateHolder(ctx context.Context, organizationID, id uuid.UUID, payload *mmodel.UpdateHolderInput, fieldsToRemove []string) (*mmodel.Holder, error) {
 	_, tracer, reqId, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -202,48 +150,7 @@ func (handler *HolderHandler) updateHolder(ctx context.Context, organizationID, 
 	return holder, nil
 }
 
-// UpdateHolder is a method that updates Holder information.
-func (handler *HolderHandler) UpdateHolder(p any, c fiber.Ctx) error {
-	ctx := c.Context()
-
-	logger, tracer, _, _ := libObservability.NewTrackingFromContext(ctx)
-
-	ctx, span := tracer.Start(ctx, "handler.update_holder_fiber")
-	defer span.End()
-
-	id, err := http.GetUUIDFromLocals(c, "id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	payload, ok := p.(*mmodel.UpdateHolderInput)
-	if !ok || payload == nil {
-		return http.WithError(c, pkg.ValidateInternalError(nil, cn.EntityHolder))
-	}
-
-	fieldsToRemove, ok := c.Locals("patchRemove").([]string)
-	if !ok {
-		libOpentelemetry.HandleSpanError(span, "Failed to get fields to remove", cn.ErrInternalServer)
-
-		logger.Log(ctx, libLog.LevelError, "Failed to get fields to remove")
-
-		return http.WithError(c, pkg.ValidateInternalError(nil, cn.EntityHolder))
-	}
-
-	holder, err := handler.updateHolder(ctx, organizationID, id, payload, fieldsToRemove)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, holder)
-}
-
-// deleteHolder is the transport-agnostic core for the holder delete.
+// deleteHolder removes a holder, soft by default and physically when hardDelete.
 func (handler *HolderHandler) deleteHolder(ctx context.Context, organizationID, id uuid.UUID, hardDelete bool) error {
 	_, tracer, reqId, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -266,33 +173,8 @@ func (handler *HolderHandler) deleteHolder(ctx context.Context, organizationID, 
 	return nil
 }
 
-// DeleteHolderByID is a method that removes Holder information by a given id.
-func (handler *HolderHandler) DeleteHolderByID(c fiber.Ctx) error {
-	ctx := c.Context()
-
-	id, err := http.GetUUIDFromLocals(c, "id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	hardDelete := http.GetBooleanParam(c, "hard_delete")
-
-	if err := handler.deleteHolder(ctx, organizationID, id, hardDelete); err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.NoContent(c)
-}
-
-// getAllHolders is the transport-agnostic core for the holder list. queries is the
-// map[string]string the caller derived from its transport (Fiber c.Queries() or the
-// Huma raw-query rebuild); http.ValidateParameters is the sole query binder so the
-// two transports validate identically.
+// getAllHolders lists holders for an organization. queries is the raw query the
+// handler rebuilt as a map; http.ValidateParameters is the sole query binder.
 func (handler *HolderHandler) getAllHolders(ctx context.Context, organizationID uuid.UUID, queries map[string]string, includeDeleted bool) (http.Pagination, error) {
 	_, tracer, reqId, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -330,23 +212,4 @@ func (handler *HolderHandler) getAllHolders(ctx context.Context, organizationID 
 	pagination.SetItems(holders)
 
 	return pagination, nil
-}
-
-// GetAllHolders retrieves Holder details by a given id
-func (handler *HolderHandler) GetAllHolders(c fiber.Ctx) error {
-	ctx := c.Context()
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	includeDeleted := http.GetBooleanParam(c, "include_deleted")
-
-	pagination, err := handler.getAllHolders(ctx, organizationID, c.Queries(), includeDeleted)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, pagination)
 }
