+++ components/ledger/internal/adapters/http/in/holder_handler.go
@@ -28,16 +28,14 @@ import (
 //     + ParseUUIDPathParameters("holder")) attached BEFORE the Huma terminal.
 //  2. Holders are ORG-SCOPED (no ledger in the path), so the shells resolve only
 //     organization_id (+ id) via the shared parsePathUUID; there is no parseOrgLedger.
-//  3. POST carries an idempotency dance: the shell resolves the client key + TTL
-//     from request headers and sets the X-Idempotency-Replayed response header from
-//     the replayed flag the shared createHolder core returns — the transport-neutral
-//     mirror of the Fiber wrapper's GetIdempotencyKeyAndTTL + c.Set.
-//  4. PATCH is RFC 7396 merge-patch: the shell derives fieldsToRemove from the parsed
-//     body via http.FindNilFields (the same derivation http.WithBody feeds into the
-//     Fiber patchRemove local) and passes it to the shared updateHolder core.
-//  5. GET-by-id, list and delete read include_deleted / hard_delete from the query,
-//     matching the Fiber http.GetBooleanParam("...") reads (ValidateParameters does
-//     not bind those flags, so the shells read them from the raw query directly).
+//  3. POST carries an idempotency dance: the handler resolves the client key + TTL
+//     from request headers and projects the replayed flag the shared createHolder
+//     core returns onto the X-Idempotency-Replayed response header.
+//  4. PATCH is RFC 7396 merge-patch: the handler derives fieldsToRemove from the parsed
+//     body via http.FindNilFields and passes it to the shared updateHolder core.
+//  5. GET-by-id, list and delete read include_deleted / hard_delete from the query;
+//     ValidateParameters does not bind those flags, so the handlers read them from
+//     the raw query directly.
 //  6. Body ops carry RawBody + SkipValidateBody so http.DecodeAndValidate is the
 //     sole body validator (never a native Huma 422). Errors go through pkgHTTP.HumaProblem.
 
@@ -48,15 +46,14 @@ var secHolderBearer = []map[string][]string{
 	{"BearerAuth": {}},
 }
 
-// queryBool mirrors http.GetBooleanParam: a query value equal to "true"
-// (case-insensitive) is true, everything else (absent, empty, "false") is false.
+// queryBool reads a boolean query flag: a value equal to "true" (case-insensitive)
+// is true, everything else (absent, empty, "false") is false.
 func queryBool(q url.Values, key string) bool {
 	return strings.EqualFold(q.Get(key), "true")
 }
 
 // queriesFromValues rebuilds the map[string]string that http.ValidateParameters
-// consumes, matching Fiber's c.Queries() (last value wins for a repeated key,
-// present-but-empty keys included).
+// consumes: last value wins for a repeated key, present-but-empty keys included.
 func queriesFromValues(q url.Values) map[string]string {
 	out := make(map[string]string, len(q))
 	for k, vs := range q {
@@ -73,28 +70,28 @@ func queriesFromValues(q url.Values) map[string]string {
 
 // --- POST /holders ------------------------------------------------------------
 
-// CreateHolderInputHuma is the Huma request envelope for POST. RawBody keeps the
-// body out of Huma's validator; the idempotency headers are read so the shell can
-// run the same claim the Fiber wrapper does.
-type CreateHolderInputHuma struct {
+// CreateHolderRequest is the Huma request envelope for POST. RawBody keeps the
+// body out of Huma's validator; the idempotency headers feed the claim the shared
+// createHolder core runs.
+type CreateHolderRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	IdempotencyKey string `header:"X-Idempotency" doc:"Idempotency key to safely retry the create; an identical retry returns the original holder"`
 	IdempotencyTTL string `header:"X-TTL" doc:"Idempotency slot TTL in seconds (default 300)"`
 	RawBody        []byte `contentType:"application/json"`
 }
 
-// CreateHolderOutputHuma pins 201 (matching http.Created) and carries the
-// X-Idempotency-Replayed response header (parity with the Fiber c.Set).
-type CreateHolderOutputHuma struct {
+// CreateHolderResponse pins 201 and carries the X-Idempotency-Replayed response
+// header.
+type CreateHolderResponse struct {
 	Status              int
 	IdempotencyReplayed string `header:"X-Idempotency-Replayed"`
 	Body                *mmodel.Holder
 }
 
-// CreateHolderHuma decodes+validates the raw body imperatively then delegates to the
+// CreateHolder decodes+validates the raw body imperatively then delegates to the
 // shared createHolder core, resolving the idempotency key/TTL from headers and
 // projecting the replayed flag onto the X-Idempotency-Replayed response header.
-func (handler *HolderHandler) CreateHolderHuma(ctx context.Context, in *CreateHolderInputHuma) (*CreateHolderOutputHuma, error) {
+func (handler *HolderHandler) CreateHolder(ctx context.Context, in *CreateHolderRequest) (*CreateHolderResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -117,28 +114,27 @@ func (handler *HolderHandler) CreateHolderHuma(ctx context.Context, in *CreateHo
 		replayedHeader = "true"
 	}
 
-	return &CreateHolderOutputHuma{Status: http.StatusCreated, IdempotencyReplayed: replayedHeader, Body: holder}, nil
+	return &CreateHolderResponse{Status: http.StatusCreated, IdempotencyReplayed: replayedHeader, Body: holder}, nil
 }
 
 // --- GET /holders/{id} --------------------------------------------------------
 
-// GetHolderInputHuma is the by-id request envelope. The path params carry no format
-// tag (ParseUUIDPathParameters is the sole validator); include_deleted is read via
-// Resolve, matching the Fiber http.GetBooleanParam read.
-type GetHolderInputHuma struct {
+// GetHolderRequest is the by-id request envelope. The path params carry no format
+// tag — ParseUUIDPathParameters is the sole validator.
+type GetHolderRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	ID             string `path:"id" doc:"Holder ID (UUID)"`
 	IncludeDeleted string `query:"include_deleted" doc:"Returns the holder even if it was logically deleted (true,false)"`
 }
 
-// GetHolderOutputHuma carries the holder verbatim.
-type GetHolderOutputHuma struct {
+// GetHolderResponse carries the holder verbatim.
+type GetHolderResponse struct {
 	Status int
 	Body   *mmodel.Holder
 }
 
-// GetHolderByIDHuma delegates to getHolderByID.
-func (handler *HolderHandler) GetHolderByIDHuma(ctx context.Context, in *GetHolderInputHuma) (*GetHolderOutputHuma, error) {
+// GetHolderByID delegates to getHolderByID.
+func (handler *HolderHandler) GetHolderByID(ctx context.Context, in *GetHolderRequest) (*GetHolderResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -154,29 +150,29 @@ func (handler *HolderHandler) GetHolderByIDHuma(ctx context.Context, in *GetHold
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &GetHolderOutputHuma{Status: http.StatusOK, Body: holder}, nil
+	return &GetHolderResponse{Status: http.StatusOK, Body: holder}, nil
 }
 
 // --- PATCH /holders/{id} ------------------------------------------------------
 
-// UpdateHolderInputHuma is the update request envelope (RawBody, see Create). The
+// UpdateHolderRequest is the update request envelope (RawBody, see Create). The
 // raw body is the sole source of the RFC 7396 null-field paths derived below.
-type UpdateHolderInputHuma struct {
+type UpdateHolderRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	ID             string `path:"id" doc:"Holder ID (UUID)"`
 	RawBody        []byte `contentType:"application/json"`
 }
 
-// UpdateHolderOutputHuma carries the updated holder (200, matching http.OK).
-type UpdateHolderOutputHuma struct {
+// UpdateHolderResponse carries the updated holder (200, matching http.OK).
+type UpdateHolderResponse struct {
 	Status int
 	Body   *mmodel.Holder
 }
 
-// UpdateHolderHuma decodes+validates the raw body imperatively, derives the merge-patch
+// UpdateHolder decodes+validates the raw body imperatively, derives the merge-patch
 // null-field paths via http.FindNilFields (the same derivation http.WithBody feeds the
 // Fiber patchRemove local), then delegates to the shared updateHolder core.
-func (handler *HolderHandler) UpdateHolderHuma(ctx context.Context, in *UpdateHolderInputHuma) (*UpdateHolderOutputHuma, error) {
+func (handler *HolderHandler) UpdateHolder(ctx context.Context, in *UpdateHolderRequest) (*UpdateHolderResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -199,25 +195,25 @@ func (handler *HolderHandler) UpdateHolderHuma(ctx context.Context, in *UpdateHo
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &UpdateHolderOutputHuma{Status: http.StatusOK, Body: holder}, nil
+	return &UpdateHolderResponse{Status: http.StatusOK, Body: holder}, nil
 }
 
 // --- DELETE /holders/{id} -----------------------------------------------------
 
-// DeleteHolderInputHuma is the delete request envelope; hard_delete is read from the
-// query (matching the Fiber http.GetBooleanParam read).
-type DeleteHolderInputHuma struct {
+// DeleteHolderRequest is the delete request envelope; hard_delete is read from the
+// query.
+type DeleteHolderRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	ID             string `path:"id" doc:"Holder ID (UUID)"`
 	HardDelete     string `query:"hard_delete" doc:"Use only to perform a physical deletion of the data. This action is irreversible. (true,false)"`
 }
 
-// DeleteHolderOutputHuma has NO Body field: paired with DefaultStatus 204 it makes
-// Huma emit a bodiless 204, matching the Fiber http.NoContent path.
-type DeleteHolderOutputHuma struct{}
+// DeleteHolderResponse has NO Body field: paired with DefaultStatus 204 it makes
+// Huma emit a bodiless 204.
+type DeleteHolderResponse struct{}
 
-// DeleteHolderByIDHuma delegates to deleteHolder; returns a bodiless 204 on success.
-func (handler *HolderHandler) DeleteHolderByIDHuma(ctx context.Context, in *DeleteHolderInputHuma) (*DeleteHolderOutputHuma, error) {
+// DeleteHolderByID delegates to deleteHolder; returns a bodiless 204 on success.
+func (handler *HolderHandler) DeleteHolderByID(ctx context.Context, in *DeleteHolderRequest) (*DeleteHolderResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -232,15 +228,15 @@ func (handler *HolderHandler) DeleteHolderByIDHuma(ctx context.Context, in *Dele
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &DeleteHolderOutputHuma{}, nil
+	return &DeleteHolderResponse{}, nil
 }
 
 // --- GET /holders (list) ------------------------------------------------------
 
-// ListHoldersInputHuma advertises the list query params (doc-only, no validation
+// ListHoldersRequest advertises the list query params (doc-only, no validation
 // tags) and captures the raw query via Resolve for the imperative binder + the
 // include_deleted flag read.
-type ListHoldersInputHuma struct {
+type ListHoldersRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	Metadata       string `query:"metadata" doc:"JSON string to filter holders by metadata fields"`
 	Limit          string `query:"limit" doc:"Max items per page (1-100, default 10)"`
@@ -257,21 +253,21 @@ type ListHoldersInputHuma struct {
 
 // Resolve captures the raw query before the handler. It performs NO validation and
 // NEVER returns an error — canonical rejection stays in http.ValidateParameters.
-func (in *ListHoldersInputHuma) Resolve(ctx huma.Context) []error {
+func (in *ListHoldersRequest) Resolve(ctx huma.Context) []error {
 	u := ctx.URL()
 	in.rawQuery = u.Query()
 
 	return nil
 }
 
-// ListHoldersOutputHuma carries the pagination envelope verbatim.
-type ListHoldersOutputHuma struct {
+// ListHoldersResponse carries the pagination envelope verbatim.
+type ListHoldersResponse struct {
 	Status int
 	Body   pkgHTTP.Pagination
 }
 
-// GetAllHoldersHuma binds the query imperatively then delegates to getAllHolders.
-func (handler *HolderHandler) GetAllHoldersHuma(ctx context.Context, in *ListHoldersInputHuma) (*ListHoldersOutputHuma, error) {
+// GetAllHolders binds the query imperatively then delegates to getAllHolders.
+func (handler *HolderHandler) GetAllHolders(ctx context.Context, in *ListHoldersRequest) (*ListHoldersResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -282,14 +278,14 @@ func (handler *HolderHandler) GetAllHoldersHuma(ctx context.Context, in *ListHol
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &ListHoldersOutputHuma{Status: http.StatusOK, Body: pagination}, nil
+	return &ListHoldersResponse{Status: http.StatusOK, Body: pagination}, nil
 }
 
 // --- GET /holders/{id}/accounts -----------------------------------------------
 
-// ListHolderAccountsInputHuma advertises the account-list query params (doc-only)
+// ListHolderAccountsRequest advertises the account-list query params (doc-only)
 // and captures the raw query via Resolve for the imperative binder.
-type ListHolderAccountsInputHuma struct {
+type ListHolderAccountsRequest struct {
 	OrganizationID string `path:"organization_id" doc:"Organization ID (UUID)"`
 	ID             string `path:"id" doc:"Holder ID (UUID)"`
 	Limit          string `query:"limit" doc:"Max items per page (1-100, default 10)"`
@@ -301,22 +297,22 @@ type ListHolderAccountsInputHuma struct {
 
 // Resolve captures the raw query before the handler (no validation; canonical
 // rejection stays in http.ValidateParameters).
-func (in *ListHolderAccountsInputHuma) Resolve(ctx huma.Context) []error {
+func (in *ListHolderAccountsRequest) Resolve(ctx huma.Context) []error {
 	u := ctx.URL()
 	in.rawQuery = u.Query()
 
 	return nil
 }
 
-// ListHolderAccountsOutputHuma carries the pagination envelope verbatim.
-type ListHolderAccountsOutputHuma struct {
+// ListHolderAccountsResponse carries the pagination envelope verbatim.
+type ListHolderAccountsResponse struct {
 	Status int
 	Body   pkgHTTP.Pagination
 }
 
-// GetAccountsByHolderHuma binds the query imperatively then delegates to
+// GetAccountsByHolder binds the query imperatively then delegates to
 // getAccountsByHolder.
-func (handler *HolderAccountsHandler) GetAccountsByHolderHuma(ctx context.Context, in *ListHolderAccountsInputHuma) (*ListHolderAccountsOutputHuma, error) {
+func (handler *HolderAccountsHandler) GetAccountsByHolder(ctx context.Context, in *ListHolderAccountsRequest) (*ListHolderAccountsResponse, error) {
 	orgID, err := parsePathUUID(in.OrganizationID, "organization_id")
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -332,10 +328,10 @@ func (handler *HolderAccountsHandler) GetAccountsByHolderHuma(ctx context.Contex
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &ListHolderAccountsOutputHuma{Status: http.StatusOK, Body: pagination}, nil
+	return &ListHolderAccountsResponse{Status: http.StatusOK, Body: pagination}, nil
 }
 
-// RegisterHolderRoutes registers the five migrated holder operations on the given
+// RegisterHolderRoutes registers the five holder operations on the given
 // Huma API. It is the per-file seam the unified server calls; the auth
 // ("midaz","holders",verb) + tenant + ParseUUIDPathParameters("holder") middleware
 // chain is attached on the versioned Fiber group BEFORE the Huma terminal, not here.
@@ -360,7 +356,7 @@ func RegisterHolderRoutes(api huma.API, h *HolderHandler, opSuffix string) {
 		// Body validated imperatively (http.DecodeAndValidate) — see file header.
 		SkipValidateBody: true,
 		DefaultStatus:    http.StatusCreated,
-	}, h.CreateHolderHuma)
+	}, h.CreateHolder)
 	attachTypedRequestBody[mmodel.CreateHolderInput](api, "createHolder"+opSuffix)
 
 	huma.Register(api, huma.Operation{
@@ -370,7 +366,7 @@ func RegisterHolderRoutes(api huma.API, h *HolderHandler, opSuffix string) {
 		Summary:     "Retrieve Holder details",
 		Tags:        []string{tag},
 		Security:    secHolderBearer,
-	}, h.GetHolderByIDHuma)
+	}, h.GetHolderByID)
 
 	huma.Register(api, huma.Operation{
 		OperationID:      "updateHolder" + opSuffix,
@@ -380,7 +376,7 @@ func RegisterHolderRoutes(api huma.API, h *HolderHandler, opSuffix string) {
 		Tags:             []string{tag},
 		Security:         secHolderBearer,
 		SkipValidateBody: true, // body validated imperatively — RFC 7396 merge-patch core.
-	}, h.UpdateHolderHuma)
+	}, h.UpdateHolder)
 	attachTypedRequestBody[mmodel.UpdateHolderInput](api, "updateHolder"+opSuffix)
 
 	huma.Register(api, huma.Operation{
@@ -392,7 +388,7 @@ func RegisterHolderRoutes(api huma.API, h *HolderHandler, opSuffix string) {
 		Security:    secHolderBearer,
 		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
 		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteHolderByIDHuma)
+	}, h.DeleteHolderByID)
 
 	huma.Register(api, huma.Operation{
 		OperationID: "listHolders" + opSuffix,
@@ -401,13 +397,13 @@ func RegisterHolderRoutes(api huma.API, h *HolderHandler, opSuffix string) {
 		Summary:     "List Holders",
 		Tags:        []string{tag},
 		Security:    secHolderBearer,
-	}, h.GetAllHoldersHuma)
+	}, h.GetAllHolders)
 }
 
 // RegisterHolderAccountsRoutes registers the holder-scoped account listing on the
 // given Huma API. It is a separate seam so the unified server can mount it
-// conditionally (only when the ledger account-query backing is wired, matching the
-// Fiber `if hah != nil` guard in crm_routes.go). Auth is ("midaz","holders","get")
+// conditionally, only when the ledger account-query backing is wired (the
+// `if hah != nil` guard in crm_routes.go). Auth is ("midaz","holders","get")
 // + ParseUUIDPathParameters("holder"), attached BEFORE the Huma terminal.
 //
 // opSuffix is appended to the operation ID — see crmOpSuffixV2.
@@ -419,5 +415,5 @@ func RegisterHolderAccountsRoutes(api huma.API, h *HolderAccountsHandler, opSuff
 		Summary:     "List Accounts by Holder",
 		Tags:        []string{"Holders"},
 		Security:    secHolderBearer,
-	}, h.GetAccountsByHolderHuma)
+	}, h.GetAccountsByHolder)
 }
